Add files via upload
I added the ability to increment a player's score by a specific number (or, as the case may be, decrement it by passing a negative number). I added HeartsSession, which keeps statistics for the overall session of hearts. HeartsSession has been integrated into the HeartsEngine in the assignPoints and checkVictory methods. A getter for the engine's HeartsSession has been added.